+++ csharpsrc/Tests/LeetCodeCases/No1.TwoSum/TestCase.cs
@@ -1,44 +1,48 @@
-﻿// <copyright file="ThreeSumTest.cs" company="Ocean">
+﻿// <copyright file="TestCase.cs" company="Ocean">
 // Copyright (c) Ocean. All rights reserved.
 // </copyright>
 
-namespace Tests.LeetCode.No15
+namespace Tests.LeetCodeCases.No1.TwoSum
 {
     using System;
-    using System.Collections.Generic;
     using System.Linq;
     using BenchmarkDotNet.Attributes;
-    using BenchmarkDotNet.Running;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using Questions.LeetCode.No15.ThreeSum;
+    using Questions.LeetCode.No1.TwoSum;
 
     /// <summary>
-    /// The test class of LeetCode No.15 question: Three Sum.
+    /// The test class of LeetCode No.1 question: Two Sum.
     /// </summary>
     [TestClass]
-    public class ThreeSumTest : LeetCodeBaseTestCase<IList<IList<int>>>
+    public class TestCase : LeetCodeBaseTestCase<int[]>
     {
         /// <summary>
-        /// The test numbers.
+        /// The test number array.
         /// </summary>
         private readonly int[] testNums;
 
         /// <summary>
-        /// The solution of LeetCode No.15 question: Three Sum.
+        /// The test target.
+        /// </summary>
+        private readonly int testTarget;
+
+        /// <summary>
+        /// The solution of LeetCode No.1 question.
         /// </summary>
         private readonly IQuestion solution;
 
         /// <summary>
-        /// The official answer of LeetCode No.15 question: Three Sum.
+        /// The official answer of No.1 question.
         /// </summary>
         private readonly IQuestion officialAnswer;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="ThreeSumTest"/> class.
+        /// Initializes a new instance of the <see cref="TestCase"/> class.
         /// </summary>
-        public ThreeSumTest()
+        public TestCase()
         {
             var random = new Random();
+            this.testTarget = random.Next(Constraints.MinValue, Constraints.MaxValue);
             this.testNums = new int[random.Next(Constraints.MinLength, Constraints.MaxLength)];
 
             for (var index = 0; index < this.testNums.Length; index++)
@@ -51,41 +55,42 @@ namespace Tests.LeetCode.No15
         }
 
         /// <summary>
-        /// Verifies the result from ThreeSum solution whether it is as same as the result from official answer.
+        /// Verifies the result from TwoSum solution whether it is as same as the result from official answer.
         /// </summary>
         [TestMethod]
-        public void VerifyThreeSumSolutionResult()
+        public void VerifyTwoSumSolutionResult()
         {
             var expectedResult = this.GetResultFromOfficialAnswer();
             var actualResult = this.GetResultFromSolution();
+            Array.Sort(expectedResult);
+            Array.Sort(actualResult);
 
-            Assert.AreEqual(expectedResult.Count, actualResult.Count);
-            Assert.IsTrue(actualResult.All(combination => combination.Sum() == 0));
+            Assert.IsTrue(actualResult.SequenceEqual(actualResult));
         }
 
         /// <summary>
-        /// Checks the performance of ThreeSum solution.
+        /// Checks the performance of TwoSum solution.
         /// </summary>
         [TestMethod]
-        public void CheckThreeSumSolutionPerformance()
+        public void CheckTwoSumSolutionPerformance()
         {
 #if RELEASE
-            BenchmarkRunner.Run<ThreeSumTest>();
+            BenchmarkRunner.Run<TwoSumTest>();
 #endif
         }
 
         /// <inheritdoc />
         [Benchmark]
-        public override IList<IList<int>> GetResultFromSolution()
+        public override int[] GetResultFromSolution()
         {
-            return this.solution.ThreeSum(this.testNums);
+            return this.solution.TwoSum(this.testNums, this.testTarget);
         }
 
         /// <inheritdoc />
         [Benchmark]
-        public override IList<IList<int>> GetResultFromOfficialAnswer()
+        public override int[] GetResultFromOfficialAnswer()
         {
-            return this.officialAnswer.ThreeSum(this.testNums);
+            return this.officialAnswer.TwoSum(this.testNums, this.testTarget);
         }
     }
 }