+++ csharpsrc/Tests/LeetCodeCases/No9.PalindromeNumber/Constraints.cs
@@ -2,12 +2,12 @@
 // Copyright (c) Ocean. All rights reserved.
 // </copyright>
 
-namespace Tests.LeetCode.No9
+namespace Tests.LeetCodeCases.No9.PalindromeNumber
 {
     /// <summary>
     /// The constraints of LeetCode No.9 question.
     /// </summary>
-    internal class Constraints
+    internal static class Constraints
     {
         /// <summary>
         /// The min value of test number.